Refactor tools project to work with Visual Studio 2017 Update 15.3+ (#2799)
In this commit, the tools project has been updated to work with newer
version of Visual Studio 2017. Nuget projects have also been updated
to utilize the new 0.2.0 version of nugetizer.

An issue addressed in this update is NuGet/Home#5071. This bug prevents
packaging from working on a hosted Visual Studio installation. Certain
project dependencies are not imported correctly and a work around in each
nuproj file had to be done.

Nugetizer also wanted to create duplicate binaries in the packages of an
AnyCPU build. This also has a work around disabling the <Pack> metadata
on particular projects.